fix(resolvers): move resolvers `tini` argument to entrypoint
Since the resolvers binary is an argument to the tini init process, tini
must be the entrypoint. However if the TektonConfig has pipelines
performance settings such as `threads-per-controller`, these values
are set as the `args` field on the container, overriding the image's
configured `CMD` array. In order to ensure the `resolver` binary is
included in the arguments to tini and not overwritten by the operator it
must be in the entrypoint.